\name{adj2mat}
\alias{adj2mat}
\title{Convert adjacency list into an adjacency matrix.}
\description{
Converts an adjacency-like list (which may or may not
contain all the gene IDs in the network) into an adjacency
matrix. This function is originally from ENA R package and the pathDESeq package uses this as an internal function for the \code{neibMat} function.
}

\author{
Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}
}
\seealso{\code{\link{neibMat}}}
\references{
Jeffrey, D. A., & Guanghua, X. (2014). ENA:Ensemble Network Aggregation  R package version 1.3-0.
}
