\name{data1065}
\alias{data1065}
\docType{data}
\title{
Simulated RNA-Seq data
}
\description{
A data frame which consists of simulated RNA-Seq gene expression data for 1065 genes.
}
\usage{data("data1065")}
\format{
  A data frame with 1065 observations for 21 variables.
  \describe{
First column consists gene names. The next 10 columns correspond to the control group and the last 10 columns correspond to the treatment group. Each row represents gene expression data for a particular gene.
  }
}
\details{
This data set was generated based on a Poisson-Gamma-Beta Markov Random Field model for given parameters in order to use as function inputs.
}
