\name{Reactome.genes}
\alias{Reactome.genes}
\docType{data}
\title{
Reactome pathway gene list 
}
\description{
This vector contains 7356 unique gene names in 24 Human Reactome pathways retrieved from Reactome pathway database 2014 version. 
}
\usage{data("Reactome.genes")}

\source{
Reactome pathway database
}
\references{
Croft, D., Mundo, A.F., Haw, R., Milacic, M., Weiser, J., Wu, G., Caudy, M., Garapati, P., Gillespie, M., Kamdar, M.R., et al. (2014). The reactome pathway knowledgebase. Nucleic Acids Res.42:D472-D477}
\keyword{datasets}
