\name{Reactome24.Hs}
\alias{Reactome24.Hs}
\docType{data}
\title{
Reactome pathway information for Homo sapiens
}
\description{
This contains information about 24 Human Reactome pathways and identified genes in each pathway. 
}
\usage{data("Reactome24.Hs")}
\format{
    \describe{ A data frame with 24 rows, each represents a Reactome pathway name and its gene list.
    
  }
}
\source{
Reactome pathway database
}
\references{
Croft, D., Mundo, A.F., Haw, R., Milacic, M., Weiser, J., Wu, G., Caudy, M., Garapati, P., Gillespie, M., Kamdar, M.R., et al.(2014). The reactome pathway knowledgebase. Nucleic Acids Res.42:D472-D477}
\keyword{datasets}
