\name{Reactome6.Hs}
\alias{Reactome6.Hs}
\docType{data}
\title{
Reactome pathway genes list for 6 selected Homo sapiens pathways
}
\description{
A vector which contains Reactome pathway gene names for 6 selected pathways. They are, Apoptosis,  Binding and Uptake of Ligands by Scavenger Receptors,
Cell Cycle,  Cell-Cell communication, Cellular responses to stress  and   Chromatin organization pathways.                                                                    
}
\usage{data("Reactome6.Hs")}

\source{
Reactome pathway database
http://www.reactome.org/pages/download-data/
}
\references{
Croft, D., Mundo A.F., Haw R., Milacic M., Weiser J., Wu G., Caudy M., Garapati P., Gillespie M., Kamdar M.R., et al.(2014). The reactome pathway knowledgebase. Nucleic Acids Res.42:D472-D477
}
