\name{symmetricize}
\alias{symmetricize}
\title{Make a matrix symmetric}
\description{
Make the matrix symmetric by making all "mirrored"
positions consistent. A variety of methods are provided to
make the matrix symmetrical. This function is originally from the ENA R package and the pathDESeq package uses this as an internal function for the \code{neibMat} function.

}

\author{
Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}
}
\seealso{\code{\link{neibMat}}}
\references{
Jeffrey, D.A., & Guanghua,X.(2014). ENA:Ensemble Network Aggregation. R package version 1.3-0.
}

