\name{ur.mat}
\alias{ur.mat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate the log likelihood using Gaussian quadrature method for up regulated genes}
\description{
This is an internal function in \code{estDE} function to calculate log likelihood values  for up regulated genes in matrix format using Gaussian quadrature. 
}
%\usage{
%ur.mat(lambda, delta, y1, y2)
%}
\seealso{\code{\link{estDE}}}
\author{
Malathi S.I. Dona
}
